+++ apps/sim/lib/copilot-new/tools/client/types.ts
@@ -1,0 +1,29 @@
+import type { BaseClientToolMetadata } from '@/lib/copilot-new/tools/client/base-tool'
+import { ClientToolCallState } from '@/lib/copilot-new/tools/client/base-tool'
+
+export interface ToolExecutionContext {
+  toolCallId: string
+  toolName: string
+  // Logging only; tools must not mutate store state directly
+  log: (level: 'debug' | 'info' | 'warn' | 'error', message: string, extra?: Record<string, any>) => void
+}
+
+export interface ToolRunResult {
+  status: number
+  message?: any
+  data?: any
+}
+
+export interface ClientToolDefinition<Args = any> {
+  name: string
+  metadata?: BaseClientToolMetadata
+  // Return true if this tool requires user confirmation before execution
+  hasInterrupt?: boolean | ((args?: Args) => boolean)
+  // Main execution entry point. Returns a result for the store to handle.
+  execute: (ctx: ToolExecutionContext, args?: Args) => Promise<ToolRunResult | void>
+  // Optional accept/reject handlers for interrupt flows
+  accept?: (ctx: ToolExecutionContext, args?: Args) => Promise<ToolRunResult | void>
+  reject?: (ctx: ToolExecutionContext, args?: Args) => Promise<ToolRunResult | void>
+}
+
+export { ClientToolCallState }
